[API]
- Se añade seguridad
- Se añade función de seguridad por token
- Se añade constante PRIVATE_API

+++ src/rutas/routes.php
@@ -28,29 +28,40 @@
  use Psr\Http\Message\ResponseInterface as Response;
 
  $app = new \Slim\App();
+ require __DIR__ . '../../functions.php';
 
   /**
 	* Descripción	: Get todos los players
 	* @author		  : <sebastian.carroza@gmail.cl> - 14/08/2019
 	* @return     JSON
 	*/
-  $app->get('/player', function(Request $request, Response $response){
-    $sql ="SELECT * FROM plaayer";
-    try{
-      $db = new db();
-      $db = $db->connectDB();
-      $resultado = $db->query($sql);
-      if ($resultado->rowCount() > 0) {
-        $players = $resultado->fetchAll(PDO::FETCH_OBJ);
-        return  $response->withJson($players);
-      }else{
-        $result = array("mensaje" => "No existen players en la DB.");
+  $app->get('/{gl_token}/player', function(Request $request, Response $response){
+    $perm = 1;
+    if(PRIVATE_API == 1){
+      $gl_token = $request->getAttribute('gl_token');
+      $perm = getStatusKey($gl_token);
+    }
+    if($perm == 1){
+      $sql ="SELECT * FROM player";
+      try{
+        $db = new db();
+        $db = $db->connectDB();
+        $resultado = $db->query($sql);
+        if ($resultado->rowCount() > 0) {
+          $players = $resultado->fetchAll(PDO::FETCH_OBJ);
+          return  $response->withJson($players);
+        }else{
+          $result = array("mensaje" => "No existen players en la DB.");
+          return  $response->withJson($result);
+        }
+        $resultado = null;
+        $db = null;
+      }catch(PDOException $e){
+        $result = array("error" => "TEXT: ".$e->getMessage());
         return  $response->withJson($result);
       }
-      $resultado = null;
-      $db = null;
-    }catch(PDOException $e){
-      $result = array("error" => "TEXT: ".$e->getMessage());
+    }else{
+      $result = array("mensaje" => "No tienes permisos para ocupar esta api");
       return  $response->withJson($result);
     }
   });
@@ -60,7 +71,13 @@
 	* @author		  : <sebastian.carroza@gmail.cl> - 14/08/2019
 	* @return     JSON
 	*/
-  $app->get('/player/{codigo}', function(Request $request, Response $response){
+  $app->get('/{gl_token}/player/{codigo}', function(Request $request, Response $response){
+    $perm = 1;
+    if(PRIVATE_API == 1){
+      $gl_token = $request->getAttribute('gl_token');
+      $perm = getStatusKey($gl_token);
+    }
+    if($perm == 1){
       $codigo = $request->getAttribute('codigo');
       $sql = "SELECT * FROM player WHERE codigo = $codigo ";
       try{
@@ -80,6 +97,10 @@
         $result = array("error" => "TEXT: ".$e->getMessage());
         return  $response->withJson($result);
       }
+    }else{
+      $result = array("mensaje" => "No tienes permisos para ocupar esta api");
+      return  $response->withJson($result);
+    }
   });
 
   /**
@@ -87,28 +108,38 @@
 	* @author		  : <sebastian.carroza@gmail.cl> - 14/08/2019
 	* @return     JSON
 	*/
-  $app->post('/player/nuevo', function(Request $request, Response $response){
-    $codigo = $request->getParam('codigo');
-    $nombre = $request->getParam('nombre');
-    $equipo = $request->getParam('equipo');
-    $sql = "INSERT INTO player (codigo, nombre, equipo) VALUES (:codigo, :nombre, :equipo) ";
-    try{
-      $db = new db();
-      $db = $db->connectDB();
-      $resultado = $db->prepare($sql);
-      $resultado->bindParam(':codigo', $codigo);
-      $resultado->bindParam(':nombre', $nombre);
-      $resultado->bindParam(':equipo', $equipo);
+  $app->post('/{gl_token}/player/nuevo', function(Request $request, Response $response){
+    $perm = 1;
+    if(PRIVATE_API == 1){
+      $gl_token = $request->getAttribute('gl_token');
+      $perm = getStatusKey($gl_token);
+    }
+    if($perm == 1){
+      $codigo = $request->getParam('codigo');
+      $nombre = $request->getParam('nombre');
+      $equipo = $request->getParam('equipo');
+      $sql = "INSERT INTO player (codigo, nombre, equipo) VALUES (:codigo, :nombre, :equipo) ";
+      try{
+        $db = new db();
+        $db = $db->connectDB();
+        $resultado = $db->prepare($sql);
+        $resultado->bindParam(':codigo', $codigo);
+        $resultado->bindParam(':nombre', $nombre);
+        $resultado->bindParam(':equipo', $equipo);
 
-      $resultado->execute();
+        $resultado->execute();
 
-      $result = array("mensaje" => "Nuevo Player guardado.");
-      return  $response->withJson($result);
+        $result = array("mensaje" => "Nuevo Player guardado.");
+        return  $response->withJson($result);
 
-      $resultado = null;
-      $db = null;
-    }catch(PDOException $e){
-      $result = array("error" => "TEXT: ".$e->getMessage());
+        $resultado = null;
+        $db = null;
+      }catch(PDOException $e){
+        $result = array("error" => "TEXT: ".$e->getMessage());
+        return  $response->withJson($result);
+      }
+    }else{
+      $result = array("mensaje" => "No tienes permisos para ocupar esta api");
       return  $response->withJson($result);
     }
   });
@@ -119,34 +150,44 @@
 	* @author		  : <sebastian.carroza@gmail.cl> - 14/08/2019
 	* @return     JSON
 	*/
-    $app->put('/player/modificar/{id}', function(Request $request, Response $response){
-      $id_player = $request->getAttribute('id');
-      $codigo = $request->getParam('codigo');
-      $nombre = $request->getParam('nombre');
-      $equipo = $request->getParam('equipo');
-      $sql = "UPDATE player
-              SET codigo = :codigo,
-                  nombre = :nombre,
-                  equipo = :equipo
-              WHERE id_player = $id_player";
-      try{
-        $db = new db();
-        $db = $db->connectDB();
-        $resultado = $db->prepare($sql);
+    $app->put('/{gl_token}/player/modificar/{id}', function(Request $request, Response $response){
+      $perm = 1;
+      if(PRIVATE_API == 1){
+        $gl_token = $request->getAttribute('gl_token');
+        $perm = getStatusKey($gl_token);
+      }
+      if($perm == 1){
+        $id_player = $request->getAttribute('id');
+        $codigo = $request->getParam('codigo');
+        $nombre = $request->getParam('nombre');
+        $equipo = $request->getParam('equipo');
+        $sql = "UPDATE player
+                SET codigo = :codigo,
+                    nombre = :nombre,
+                    equipo = :equipo
+                WHERE id_player = $id_player";
+        try{
+          $db = new db();
+          $db = $db->connectDB();
+          $resultado = $db->prepare($sql);
 
-        $resultado->bindParam(':codigo', $codigo);
-        $resultado->bindParam(':nombre', $nombre);
-        $resultado->bindParam(':equipo', $equipo);
+          $resultado->bindParam(':codigo', $codigo);
+          $resultado->bindParam(':nombre', $nombre);
+          $resultado->bindParam(':equipo', $equipo);
 
-        $resultado->execute();
+          $resultado->execute();
 
-        $result = array("mensaje" => "Player modificado.");
-        return  $response->withJson($result);
+          $result = array("mensaje" => "Player modificado.");
+          return  $response->withJson($result);
 
-        $resultado = null;
-        $db = null;
-      }catch(PDOException $e){
-        $result = array("error" => "TEXT: ".$e->getMessage());
+          $resultado = null;
+          $db = null;
+        }catch(PDOException $e){
+          $result = array("error" => "TEXT: ".$e->getMessage());
+          return  $response->withJson($result);
+        }
+      }else{
+        $result = array("mensaje" => "No tienes permisos para ocupar esta api");
         return  $response->withJson($result);
       }
     });
@@ -156,28 +197,38 @@
   	* @author		  : <sebastian.carroza@gmail.cl> - 14/08/2019
   	* @return     JSON
   	*/
-      $app->delete('/player/delete/{id}', function(Request $request, Response $response){
-        $id_player = $request->getAttribute('id');
-        $sql = "DELETE FROM player
-                WHERE id_player = $id_player";
-        try{
-          $db = new db();
-          $db = $db->connectDB();
-          $resultado = $db->prepare($sql);
-          $resultado->execute();
+      $app->delete('/{gl_token}/player/delete/{id}', function(Request $request, Response $response){
+        $perm = 1;
+        if(PRIVATE_API == 1){
+          $gl_token = $request->getAttribute('gl_token');
+          $perm = getStatusKey($gl_token);
+        }
+        if($perm == 1){
+          $id_player = $request->getAttribute('id');
+          $sql = "DELETE FROM player
+                  WHERE id_player = $id_player";
+          try{
+            $db = new db();
+            $db = $db->connectDB();
+            $resultado = $db->prepare($sql);
+            $resultado->execute();
 
-          if ($resultado->rowCount() > 0) {
-            $result = array("mensaje" => "Player eliminado.");
-            return  $response->withJson($result);
-          }else{
-            $result = array("mensaje" => "No existe Player con esa id.");
+            if ($resultado->rowCount() > 0) {
+              $result = array("mensaje" => "Player eliminado.");
+              return  $response->withJson($result);
+            }else{
+              $result = array("mensaje" => "No existe Player con esa id.");
+              return  $response->withJson($result);
+            }
+
+            $resultado = null;
+            $db = null;
+          }catch(PDOException $e){
+            $result = array("error" => "TEXT: ".$e->getMessage());
             return  $response->withJson($result);
           }
-
-          $resultado = null;
-          $db = null;
-        }catch(PDOException $e){
-          $result = array("error" => "TEXT: ".$e->getMessage());
+        }else{
+          $result = array("mensaje" => "No tienes permisos para ocupar esta api");
           return  $response->withJson($result);
         }
       });
@@ -199,9 +250,9 @@
     	* @author		  : <sebastian.carroza@gmail.cl> - 14/08/2019
     	* @return     JSON
     	*/
-      $app->get('/statuskey/{gl_key}', function(Request $request, Response $response){
-        $gl_key = $request->getAttribute('gl_key');
-        $sql = "SELECT bo_estado FROM app WHERE gl_key = '$gl_key' ";
+      $app->get('/statuskey/{gl_token}', function(Request $request, Response $response){
+        $gl_token = $request->getAttribute('gl_token');
+        $sql = "SELECT bo_estado FROM app WHERE gl_token = '$gl_token' ";
         try{
           $db = new db();
           $db = $db->connectDB();
